+++ Classes/Domain/Model/AbstractModel.php
@@ -9,6 +9,9 @@ use Illuminate\Database\Eloquent\Model;
  * @package T3v\T3vDataMapper\Domain\Model
  */
 abstract class AbstractModel extends Model {
+  /**
+   * The table identifier for `created_at` and `updated_at`.
+   */
   const CREATED_AT = 'crdate';
   const UPDATED_AT = 'tstamp';
 
@@ -20,4 +23,13 @@ abstract class AbstractModel extends Model {
    * @var string
    */
   protected $primaryKey = 'uid';
+
+  /**
+   * The date format.
+   *
+   * Overwrites the default storage format of the model's date columns.
+   *
+   * @var string
+   */
+  protected $dateFormat = 'U';
 }