+++ CHANGELOG.md
@@ -3,8 +3,8 @@ CHANGELOG
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
-Unreleased
-----------
+7.2.0
+-----
 
 * Improved `AbstractModel` and `PageLanguageOverlayCommandController`
 * Introduced `strict` and `fallback` mode (breaking change)